apps/api/pyproject.toml
[project]
name = "ideaforge"
version = "0.1.0"
description = "IdeaForge — multi-agent business idea research platform"
requires-python = ">=3.11"
dependencies = [
    # Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    # Agents (without [tools] extras to avoid onnxruntime platform issues)
    "crewai>=0.100.0",
    # Vector DB (HTTP client only — connects to ChromaDB Docker container)
    "chromadb-client>=0.5.0",
    # Data sources
    "praw>=7.8.0",
    "google-api-python-client>=2.155.0",
    "httpx>=0.28.0",
    # Trends
    "pytrends>=4.9.0",
    # Sentiment
    "vaderSentiment>=3.3.2",
    # Config
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",
    # Logging
    "structlog>=24.4.0",
    # Utils
    "pyyaml>=6.0.0",
]

[tool.uv]
override-dependencies = [
    # onnxruntime 1.24+ dropped macOS x86_64 wheels
    "onnxruntime>=1.19.0,<1.24",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
